pyproject.toml
[project]
name = "cyber-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "logfire>=4.24.0",
    "pydantic-ai>=1.59.0",
    "python-dotenv>=1.2.1",
]
